+++ CHANGELOG.md
@@ -51,3 +51,4 @@
 - [x] plugin fault-injection [fault-injection](https://apisix.apache.org/zh/docs/apisix/plugins/fault-injection/) 2024-04-19
 - [x] plugin redirect [redirect](https://apisix.apache.org/zh/docs/apisix/plugins/redirect/) 2024-04-20
 - [x] plugin csrf [csrf](https://apisix.apache.org/zh/docs/apisix/plugins/csrf/) 2024-04-21
+- [x] plugin prometheus [prometheus](https://apisix.apache.org/zh/docs/apisix/plugins/prometheus/) 2024-04-21